% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon_data.R
\name{pokemon_data}
\alias{pokemon_data}
\title{Show statistics of pokemons, including weight, height and strength.}
\usage{
pokemon_data()
}
\value{
A dataset containing statistics of pokemons
}
\description{
Show statistics of pokemons, including weight, height and strength.
}
