% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon_data_partial.R
\name{pokemon_data_partial}
\alias{pokemon_data_partial}
\title{Show chosen statistics of pokemons}
\usage{
pokemon_data_partial(para = "Weight")
}
\arguments{
\item{para}{specific characteristic you want to check about pokemon data, like "weight", "height", "HP"...,set default to "Weight"}
}
\value{
A dataset containing statistics of pokemons that chosen by the user
}
\description{
Show chosen statistics of pokemons
}
