% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_pokemon_data_partial.R
\name{summary_pokemon_data_partial}
\alias{summary_pokemon_data_partial}
\title{Show summary statistics of pokemons}
\usage{
summary_pokemon_data_partial(para = "Weight", summary = TRUE)
}
\arguments{
\item{para}{specific characteristic you want to check about pokemon data, like "weight", "height", "HP"..., set default to "Weight"}

\item{summary}{TURE if you want to see the summary statistics, FALSE if you do not want to see, set default to TRUE}
}
\value{
A dataset containing suammry statistics of pokemons that chosen by the user
}
\description{
Show summary statistics of pokemons
}
